apps/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
pydantic
pydantic-settings
supabase==2.9.0
httpx>=0.26.0
boto3==1.42.9
botocore==1.42.9
jmespath==1.0.1
s3transfer==0.16.0
urllib3==2.6.2
python-multipart==0.0.20
aiohttp>=3.8.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
